+++ CHANGELOG.md
@@ -3,9 +3,18 @@
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
 <a name="0.1.0"></a>
-# 0.1.0 (2018-10-14)
+# 0.1.0 (2018-10-21)
 
 
 ### Features
 
 * a simple koa-server ([e2b0217](https://github.com/cytle/koagent/commit/e2b0217))
+* add certificate ([8dbc006](https://github.com/cytle/koagent/commit/8dbc006))
+* certificate-middleware ([13ae187](https://github.com/cytle/koagent/commit/13ae187))
+* eslint support typescript ([12a2535](https://github.com/cytle/koagent/commit/12a2535))
+* lerna ([a294436](https://github.com/cytle/koagent/commit/a294436))
+* proxy ([b94bb89](https://github.com/cytle/koagent/commit/b94bb89))
+* router rootCA ([c059572](https://github.com/cytle/koagent/commit/c059572))
+* start script ([7083fe4](https://github.com/cytle/koagent/commit/7083fe4))
+* ts build and lint ([461690d](https://github.com/cytle/koagent/commit/461690d))
+* use http-proxy to proxy request ([aa907e9](https://github.com/cytle/koagent/commit/aa907e9))